Fix BMX multi-machine playback and widen sample addressing to u64
- Add is_tracker flag to NodeType::BuzzMachine, use real machine names
  from BMX file instead of hardcoding "Tracker"
- Fix non-tracker tracks getting machine_node: None (now resolves name)
- Add Matilde Tracker (Mono) to is_tracker_dll recognition
- Fix base_channel computation: use running counter instead of node IDs
- Slice song.channels per-machine in init_machines via channels_for_node
- Make play_pattern track-aware: plays only selected track, silences others
- Widen position/increment from u32 to u64 (16.16 fixed-point) to support
  BMX samples longer than 65535 frames without overflow

+++ crates/mb-engine/src/channel.rs
@@ -36,10 +36,10 @@ impl ActiveMod {
 pub struct ChannelState {
     /// Current sample index
     pub sample_index: u8,
-    /// Current position in sample (16.16 fixed-point)
-    pub position: u32,
-    /// Playback increment (16.16 fixed-point)
-    pub increment: u32,
+    /// Current position in sample (16.16 fixed-point, u64 for large samples)
+    pub position: u64,
+    /// Playback increment (16.16 fixed-point, u64 for large samples)
+    pub increment: u64,
     /// Is the channel currently playing?
     pub playing: bool,
 
@@ -147,7 +147,7 @@ impl ChannelState {
         match effect {
             Effect::SetVolume(v) => self.volume = (*v).min(64),
             Effect::SetPan(p) => self.panning = (*p as i16 - 128).clamp(-64, 64) as i8,
-            Effect::SampleOffset(o) => self.position = (*o as u32) << 24,
+            Effect::SampleOffset(o) => self.position = (*o as u64) << 24,
             Effect::FineVolumeSlideUp(v) => {
                 self.volume = (self.volume as i16 + *v as i16).clamp(0, 64) as u8;
             }
@@ -346,10 +346,10 @@ impl ChannelState {
 
         // Handle looping
         let pos_samples = self.position >> 16;
-        if sample.has_loop() && pos_samples >= sample.loop_end {
-            let loop_len = sample.loop_end - sample.loop_start;
+        if sample.has_loop() && pos_samples >= sample.loop_end as u64 {
+            let loop_len = (sample.loop_end - sample.loop_start) as u64;
             self.position -= loop_len << 16;
-        } else if pos_samples >= sample.len() as u32 {
+        } else if pos_samples >= sample.len() as u64 {
             self.playing = false;
         }
